+++ packages/medusa/src/api/routes/admin/products/get-product.ts
@@ -118,7 +118,11 @@ export default async (req, res) => {
   res.json({ product })
 }
 
-async function getProductWithIsolatedProductModule(req, id, retrieveConfig) {
+export async function getProductWithIsolatedProductModule(
+  req,
+  id,
+  retrieveConfig
+) {
   // TODO: Add support for fields/expands
   const remoteQuery = req.scope.resolve("remoteQuery")