+++ packages/medusa/src/scripts/create-default-rule-types.ts
@@ -11,17 +11,28 @@ export const createDefaultRuleTypes = async (container: AwilixContainer) => {
     "pricingModuleService"
   )
   const existing = await pricingModuleService.listRuleTypes(
-    { rule_attribute: ["region_id"] },
-    { take: 1 }
+    { rule_attribute: ["region_id", "customer_group_id"] },
+    { take: 2 }
   )
 
-  if (existing.length) {
+  if (existing.length === 2) {
     return
   }
 
-  await pricingModuleService.createRuleTypes([
-    { name: "region_id", rule_attribute: "region_id" },
-  ])
+  if (existing.length === 0) {
+    await pricingModuleService.createRuleTypes([
+      { name: "region_id", rule_attribute: "region_id" },
+      { name: "customer_group_id", rule_attribute: "customer_group_id" },
+    ])
+  } else if (existing[0].rule_attribute === "region_id") {
+    await pricingModuleService.createRuleTypes([
+      { name: "customer_group_id", rule_attribute: "customer_group_id" },
+    ])
+  } else {
+    await pricingModuleService.createRuleTypes([
+      { name: "region_id", rule_attribute: "region_id" },
+    ])
+  }
 }
 
 const migrate = async function ({ directory }) {
@@ -38,7 +49,9 @@ const migrate = async function ({ directory }) {
 migrate({ directory: process.cwd() })
   .then(() => {
     console.log("Created default rule types")
+    process.exit()
   })
   .catch(() => {
     console.log("Failed to create rule types")
+    process.exit(1)
   })