+++ packages/types/src/pricing/common/price-rule.ts
@@ -6,33 +6,54 @@ import { RuleTypeDTO } from "./rule-type"
  * @interface
  * 
  * A price rule's data.
- * 
- * @prop id - The ID of the price rule.
- * @prop price_set_id - The ID of the associated price set.
- * @prop price_set - The associated price set. It may only be available if the relation `price_set` is expanded.
- * @prop rule_type_id - The ID of the associated rule type.
- * @prop rule_type - The associated rule type. It may only be available if the relation `rule_type` is expanded.
- * @prop value - The value of the price rule.
- * @prop priority - The priority of the price rule in comparison to other applicable price rules.
- * @prop price_set_money_amount_id - The ID of the associated price set money amount.
- * @prop price_list_id - The ID of the associated price list.
  */
 export interface PriceRuleDTO {
+  /**
+   * The ID of the price rule.
+   */
   id: string
+  /**
+   * The ID of the associated price set.
+   */
   price_set_id: string
+  /**
+   * The associated price set.
+   * 
+   * @expandable
+   */
   price_set: PriceSetDTO
+  /**
+   * The ID of the associated rule type.
+   */
   rule_type_id: string
+  /**
+   * The associated rule type.
+   * 
+   * @expandable
+   */
   rule_type: RuleTypeDTO
   /** 
    * @ignore
-   * @privateRemark
+   * @privateRemarks
    * 
    * Behavior behind this property is not implemented yet.
    */
   is_dynamic: boolean
+  /**
+   * The value of the price rule.
+   */
   value: string
+  /**
+   * The priority of the price rule in comparison to other applicable price rules.
+   */
   priority: number
+  /**
+   * The ID of the associated price set money amount.
+   */
   price_set_money_amount_id: string
+  /**
+   * The ID of the associated price list.
+   */
   price_list_id: string
 }
 
@@ -41,26 +62,30 @@ export interface PriceRuleDTO {
  * @interface
  * 
  * A price rule to create.
- * 
- * @prop id - The ID of the price rule.
- * @prop price_set_id - The ID of the associated price set.
- * @prop rule_type_id - The ID of the associated rule type.
- * @prop value - The value of the price rule.
- * @prop priority - The priority of the price rule in comparison to other applicable price rules.
- * @prop price_set_money_amount_id - The ID of the associated price set money amount.
- * @prop price_list_id - The ID of the associated price list.
  */
 export interface CreatePriceRuleDTO {
+  /**
+   * The ID of the price rule.
+   */
   id: string
+  /**
+   * The ID of the associated price set.
+   */
   price_set_id: string
+  /**
+   * The ID of the associated rule type.
+   */
   rule_type_id: string
   /** 
    * @ignore
-   * @privateRemark
+   * @privateRemarks
    * 
    * Behavior behind this property is not implemented yet.
    */
   is_dynamic?: boolean
+  /**
+   * The value of the price rule.
+   */
   value: string
   priority?: number
   price_set_money_amount_id: string
@@ -71,14 +96,6 @@ export interface CreatePriceRuleDTO {
  * @interface
  * 
  * The data to update in a price rule. The `id` is used to identify which money amount to update.
- * 
- * @prop id - The ID of the price rule to update.
- * @prop price_set_id - The ID of the associated price set.
- * @prop rule_type_id - The ID of the associated rule type.
- * @prop value - The value of the price rule.
- * @prop priority - The priority of the price rule in comparison to other applicable price rules.
- * @prop price_set_money_amount_id - The ID of the associated price set money amount.
- * @prop price_list_id - The ID of the associated price list.
  */
 export interface UpdatePriceRuleDTO {
   id: string
@@ -91,26 +108,45 @@ export interface UpdatePriceRuleDTO {
    * Behavior behind this property is not implemented yet.
    */
   is_dynamic?: boolean
+  /**
+   * The value of the price rule.
+   */
   value?: string
+  /**
+   * The priority of the price rule in comparison to other applicable price rules.
+   */
   priority?: number
+  /**
+   * The ID of the associated price set money amount.
+   */
   price_set_money_amount_id?: string
+  /**
+   * The ID of the associated price list.
+   */
   price_list_id?: string
 }
 
 /**
  * @interface
  * 
- * Filters to apply to price rules.
- * 
- * @prop id - The IDs to filter price rules by.
- * @prop name - The names to filter price rules by.
- * @prop price_set_id - The IDs to filter the price rule's associated price set.
- * @prop rule_type_id - The IDs to filter the price rule's associated rule type.
+ * Filters to apply on price rules.
  */
 export interface FilterablePriceRuleProps
   extends BaseFilterable<FilterablePriceRuleProps> {
+  /**
+   * The IDs to filter price rules by.
+   */
   id?: string[]
+  /**
+   * The names to filter price rules by.
+   */
   name?: string[]
+  /**
+   * The IDs to filter the price rule's associated price set.
+   */
   price_set_id?: string[]
+  /**
+   * The IDs to filter the price rule's associated rule type.
+   */
   rule_type_id?: string[]
 }